+++ lib/image_processing/mini_magick.rb
@@ -52,6 +52,12 @@ module ImageProcessing
         magick.extent "#{width}x#{height}"
       end
 
+      def limit(magick, limits)
+        limit_args = limits.flat_map { |type, value| %W[-limit #{type} #{value}] }
+        magick.args.replace limit_args + magick.args
+        magick
+      end
+
       def append(magick, *args)
         magick.merge! args
       end